+++ src/main/java/com/hubspot/jinjava/interpret/ExtendsTagCycleException.java
@@ -1,17 +1,10 @@
 package com.hubspot.jinjava.interpret;
 
-public class ExtendsTagCycleException extends InterpretException {
-  private static final long serialVersionUID = -3058494056577268723L;
-
-  private final String path;
+public class ExtendsTagCycleException extends TagCycleException {
+  private static final long serialVersionUID = 3183769038400532542L;
 
   public ExtendsTagCycleException(String path, int lineNumber) {
-    super("Extends tag cycle for path '" + path + "'", lineNumber);
-    this.path = path;
-  }
-
-  public String getPath() {
-    return path;
+    super("Extends", path, lineNumber);
   }
 
 }