+++ src/main/java/com/hubspot/jinjava/interpret/IncludeTagCycleException.java
@@ -1,0 +1,10 @@
+package com.hubspot.jinjava.interpret;
+
+public class IncludeTagCycleException extends TagCycleException {
+  private static final long serialVersionUID = -5487642459443650227L;
+
+  public IncludeTagCycleException(String path, int lineNumber) {
+    super("Include", path, lineNumber);
+  }
+
+}