+++ CHANGES.md
@@ -1,7 +1,9 @@
 # Jinjava Releases #
 
-### Version 2.0.12 ([Maven Central](http://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.hubspot.jinjava%22%20AND%20v%3A%222.0.11%22)) ###
+### Version 2.1.0 ([Maven Central](http://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.hubspot.jinjava%22%20AND%20v%3A%222.1.0%22)) ###
 
+* Refactored node render logic to return richer OutputNode instances, removing a need for a special intermediate string value in text output
+* Refactored cycle detection in import and include tags to remove use of special named vars in context
 * Fix bug to properly detect cycles in extends tag
 * Fix infinite recursion bug in resolveBlockStubs when block contains self-reference